pyproject.toml
[project]
name = "neptunium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "huggingface-hub>=1.3.2",
    "llama-cpp-python>=0.3.2",
]
